requirements.txt
flask
chromadb
openai
tiktoken
google-genai
docker
pandas